Add `useWidget` hook to make Widget component
